requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
httpx==0.28.1
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
pytest==8.4.1
alembic==1.13.2
psycopg[binary]==3.2.1
